Fully replace old button with new one
