+++ nym-vpn-app/src/contexts/autologin/PincodeDialog.tsx
@@ -68,11 +68,13 @@ export function PincodeDialog({
 
         <PinCodeDigits code={code} />
 
-        <Button className="mt-3 w-full" onClick={handleClick}>
-          <div className="flex items-center justify-center gap-2">
-            <MsIcon icon="content_copy" className="text-xl!" />
-            <span>{t('autologin.copy-code')}</span>
-          </div>
+        <Button
+          // className="mt-3 flex w-full items-center justify-center gap-2"
+          onClick={handleClick}
+          variant="primary"
+        >
+          <MsIcon icon="content_copy" className="text-xl" />
+          <span>{t('autologin.copy-code')}adfaf</span>
         </Button>
       </div>
     </Dialog>
