+++ nym-vpn-app/src/ui/index.ts
@@ -1,6 +1,5 @@
-export { default as Button } from './Button';
 export { default as ButtonIcon } from './ButtonIcon';
-export { default as ButtonNew } from './ButtonNew';
+export { default as Button } from './Button';
 export { default as ButtonText } from './ButtonText';
 export { default as Dialog } from './Dialog';
 export { default as DraggableList } from './DraggableList';
@@ -25,9 +24,8 @@ export { default as Skeleton } from './Skeleton';
 export { default as Toast } from './Toast';
 export { default as ConfirmationDialog } from './ConfirmationDialog';
 export { default as Spinner } from './Spinner';
-export * from './Button';
 export * from './ButtonIcon';
-export * from './ButtonNew';
+export * from './Button';
 export * from './ButtonText';
 export * from './Dialog';
 export * from './DraggableList';
